// frontend/src/api/model/patientPracticeRead.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Hospital Internment API
 * OpenAPI spec version: 0.1.0
 */
import type { PracticeOrderStatus } from './practiceOrderStatus';

export interface PatientPracticeRead {
  hospitalization_id: string;
  code: string;
  name: string;
  status: PracticeOrderStatus;
  /** @pattern ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$ */
  quantity: string;
  prescribed_at: string;
  performed_at: string | null;
  is_consultation: boolean;
  amount: string | null;
  indication: string | null;
}